+++ CHANGES.rst
@@ -4,10 +4,13 @@ Changelog
 0.5dev
 ------
 
-- Integrate hide_cart_if_empty property of cart data interface.
+- Integrate ``cart_count_limit`` property of cart item data interface.
   [rnix]
 
-- Implement display property of stock interface and consider it in buyable
+- Integrate ``hide_cart_if_empty`` property of cart data interface.
+  [rnix]
+
+- Implement ``display`` property of stock interface and consider it in buyable
   controls.
   [rnix]